backend/requirements.txt
asgiref==3.6.0
django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pyjwt==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7